# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TraclyTag Admin Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    delete:
      operationId: deleteUser
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
    put:
      operationId: updateUser
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    Role:
      type: string
      enum: [super_master, master, admin, client_admin, operator]
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          $ref: "#/components/schemas/Role"
        companyId:
          type: ["integer", "null"]
        companyName:
          type: ["string", "null"]
        isActive:
          type: boolean
        enabledModules:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, username, email, phone, role, companyId, companyName, isActive, enabledModules, createdAt]
    CreateUserBody:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        password:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        companyId:
          type: ["integer", "null"]
        isActive:
          type: boolean
        enabledModules:
          type: string
      required: [username, email, password, role]
    UpdateUserBody:
      type: object
      properties:
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          $ref: "#/components/schemas/Role"
        isActive:
          type: boolean
        enabledModules:
          type: string
        password:
          type: string
        companyId:
          type: ["integer", "null"]
...