# lib/api-zod/src/generated/types/role.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Api
 * TraclyTag Admin Panel API
 * OpenAPI spec version: 0.1.0
 */

export type Role = typeof Role[keyof typeof Role];


export const Role = {
  super_master: 'super_master',
  master: 'master',
  admin: 'admin',
  client_admin: 'client_admin',
  operator: 'operator',
} as const;