# lib/api-zod/src/generated/types/updateUserBody.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Api
 * TraclyTag Admin Panel API
 * OpenAPI spec version: 0.1.0
 */
import type { Role } from './role';

export interface UpdateUserBody {
  email?: string;
  /** @nullable */
  phone?: string | null;
  role?: Role;
  isActive?: boolean;
  enabledModules?: string;
  password?: string;
  /** @nullable */
  companyId?: number | null;
}